+++ docs/changes.rst
@@ -79,6 +79,10 @@ To be released.
   [:commit:`88574468a38015669dae903185fb328abdd717c0` by Taeho Kim]
 - :exc:`wand.resource.DestroyedResourceError` becomes a subtype of
    :exc:`wand.exceptions.WandException`.
+- :class:`~wand.color.Color` is now hashable, so can be used as a key of
+  dictionaries, or an element of sets.  [:issue:`114` by klutzy]
+- :class:`~wand.color.Color` has :attr:`~wand.color.Color.normalized_string`
+  property.
 
 __ http://en.wikipedia.org/wiki/Seam_carving